BGDIINF_SB-2019: moving attrib class to correct file
Moving the class for LayerAttribution to AbstractLayer file, as it is there that it is firstly described and used.
Also adding some comment in OpenLayersInternalLayer to clarify why we have specific code for WMTS external layer definition here, and why not for WMS
