+++ src/api/layers/GeoAdminVectorLayer.class.js
@@ -1,4 +1,5 @@
-import GeoAdminLayer, { LayerAttribution } from '@/api/layers/GeoAdminLayer.class'
+import { LayerAttribution } from '@/api/layers/AbstractLayer.class'
+import GeoAdminLayer from '@/api/layers/GeoAdminLayer.class'
 import LayerTypes from '@/api/layers/LayerTypes.enum'
 
 /** Metadata for a vector tile layer (MapLibre layer) served by our backend */
